Added DTOs package and moved MovieGenreDTO to DTOs
