+++ src/main/java/com/moovy/entity/Movie.java
@@ -1,12 +1,11 @@
 package com.moovy.entity;
 
-import com.moovy.MovieGenreDTO;
+import com.moovy.dtos.MovieGenreDTO;
 import jakarta.persistence.*;
 import lombok.Data;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.util.Set;
 
 @Entity
 @Data
